Make the server's CLOSEWAIT state have a timeout
The tcb for the closing socket is only released after a timewait,
which gives a chance to reply on duplicated FIN segment. Otherwise
, the client missing the FINACK segment will always retry until
exceeding time limits.

However, as the closewait timer is non-blocking to the close api,
the server may shut down while the client still attempts to get
ack segments. An error return value will be transfered to cause
the client exit in under an exceptional status.